\name{f3}
\alias{f3}
\alias{f4}
\title{
  nothing that matters. Just for testing
}
\description{
  multiply by two and extract first element
}
\usage{
f3(v1)
f4(x)
}
\arguments{
  \item{x}{
    single value
  }
  \item{v1}{
  a vector
}
}
\value{
  A single value.
}
\author{
  me
}
\examples{
f3(4:12)
f4(5)
}
\keyword{ misc}

